+++ src/CascaderPicker/props.d.ts
@@ -73,5 +73,10 @@ export interface ICascaderProps extends IBaseFormItemPropsWithOutFocus {
     selectedOptions: ICascaderOption[],
     e: Record<string, any>
   ) => void;
+  /**
+   * @description 点击蒙层是否可以关闭
+   * @default false
+   */
+  maskClosable?: boolean;
 }
 export declare const CascaderDefaultProps: Partial<ICascaderProps>;