+++ src/RangePicker/props.js
@@ -6,6 +6,7 @@ export const DateRangePickerDefaultProps = {
   splitCharacter: '-',
   startPlaceholder: '未选择',
   endPlaceholder: '未选择',
+  maskClosable: false,
   onFormat(values, valueStrs) {
     const { format, splitCharacter } = this.props;
     if (format && valueStrs && valueStrs[0] && valueStrs[1]) {